refactor: share goal time formatting
