+++ apps/kimi-code/src/tui/components/messages/goal-format.ts
@@ -1,0 +1,13 @@
+export function formatGoalElapsed(ms: number): string {
+  const totalSeconds = Math.round(ms / 1000);
+  if (totalSeconds < 60) return `${String(totalSeconds)}s`;
+  const minutes = Math.floor(totalSeconds / 60);
+  const seconds = totalSeconds % 60;
+  if (minutes < 60) return `${String(minutes)}m ${seconds.toString().padStart(2, '0')}s`;
+  const hours = Math.floor(minutes / 60);
+  return `${String(hours)}h ${(minutes % 60).toString().padStart(2, '0')}m`;
+}
+
+export function pluralizeGoalCount(n: number, singular: string, plural?: string): string {
+  return `${String(n)} ${n === 1 ? singular : (plural ?? `${singular}s`)}`;
+}
